requirements/base.txt
beautifulsoup4==4.10.0
lxml==4.8.0
requests==2.27.1
python-telegram-bot==13.11